+++ Facebook/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 6.20.0.0
+* Certified with Facebook SDK 6.20.0.
+* Removed deprecated code paths based on the minimum supported AppLovin MAX SDK version 13.0.0.
+* Updated ad display failed error code.
+
 ## 6.17.1.0
 * Certified with Facebook SDK 6.17.1.